+++ model/CHANGELOG.md
@@ -4,7 +4,7 @@ This file documents changes to the `workbench-model` library, including notes on
 
 ## 0.15
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-model" % "0.15-f9f0d4c"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-model" % "0.15-TRAVIS-REPLACE-ME"`
 
 ### Added
 - AzureB2CId to WorkbenchUser